refactor(all): use Matrix operations
also small changes on filenames and export methods.
