+++ package.json
@@ -2,7 +2,7 @@
   "name": "ml-fnn",
   "version": "2.1.1",
   "description": "feedforward neural networks library",
-  "main": "src/index.js",
+  "main": "src/FeedForwardNeuralNetwork.js",
   "directories": {
     "lib": "src",
     "test": "test"
