+++ src/activationFunctions.js
